+++ plugins/test/avatar/plugin.d
@@ -39,7 +39,7 @@ final class AvatarClient : IPlugin
 		connection.registerPacket!UpdateAvatarsPacket(&handleUpdateAvatarsPacket);
 	}
 
-	void drawEntities(ref Render1Event event)
+	void drawEntities(ref RenderSolid3dEvent event)
 	{
 		graphics.chunkShader.bind;
 		graphics.draw(batch);