+++ plugins/voxelman/block/plugin.d
@@ -32,8 +32,8 @@ public:
 	override string id() @property { return "voxelman.block.blockmanager"; }
 	override void preInit()
 	{
-		regBlock("unknown").color(0,0,0).isVisible(false).isSolid(true).meshHandler(&makeNullMesh);
-		regBlock("air").color(0,0,0).isVisible(false).isSolid(false).meshHandler(&makeNullMesh);
+		regBlock("unknown").color(0,0,0).isVisible(false).solidity(Solidity.solid).meshHandler(&makeNullMesh);
+		regBlock("air").color(0,0,0).isVisible(false).solidity(Solidity.transparent).meshHandler(&makeNullMesh);
 	}
 
 	BlockInfoSetter regBlock(string name) {
@@ -61,7 +61,7 @@ mixin template BlockPluginCommonImpl()
 		bm.regBlock("dirt").colorHex(0x835929).meshHandler(&makeColoredBlockMesh);
 		bm.regBlock("stone").colorHex(0x8B8D7A).meshHandler(&makeColoredBlockMesh);
 		bm.regBlock("sand").colorHex(0xA68117).meshHandler(&makeColoredBlockMesh);
-		bm.regBlock("water").colorHex(0x0055AA).meshHandler(&makeColoredBlockMesh);
+		bm.regBlock("water").colorHex(0x0055AA).meshHandler(&makeColoredBlockMesh).solidity(Solidity.semiTransparent);
 		registerResourcesImpl(resmanRegistry);
 	}